+++ src/pages/hyominShin/Main/Article/ArticleBody.scss
@@ -1,98 +1,5 @@
 .articleBody {
-  float: left;
   width: 65vh;
   height: 100vh;
-
-  .feeds {
-    margin-top: 100px;
-    border: 1px solid lightgray;
-    border-radius: 2px;
-
-    .articleFeedImg {
-      width: 100%;
-      height: 60vh;
-      border-bottom: 1px solid gray;
-      object-fit: cover;
-    }
-
-    .reaction {
-      padding: 0 16px;
-
-      .commentAndFeedTime {
-        display: flex;
-        flex-direction: column;
-        justify-content: space-between;
-        height: 15vh;
-
-        .commentList {
-          margin-top: 10px;
-          margin-left: 5px;
-
-          .comment {
-            .textBold {
-              font-weight: bold;
-            }
-
-            .smallHeart {
-              width: 14px;
-              margin-top: 3px;
-              margin-right: 5px;
-              float: right;
-            }
-
-            .miniSubGray {
-              margin-top: 4px;
-              margin-right: 3px;
-              float: right;
-
-              &:hover {
-                cursor: default;
-              }
-            }
-          }
-        }
-
-        .feedTimeCount {
-          margin-left: 5px;
-          color: #8e8e8e;
-          font-size: 14px;
-        }
-      }
-    }
-
-    .commentDiv {
-      position: relative;
-      height: 45px;
-      margin-top: 10px;
-      border-top: 1px solid lightgray;
-
-      .commentInput {
-        position: absolute;
-        width: 85%;
-        height: 39px;
-        margin-top: 2px;
-        margin-left: 17px;
-        border: none;
-
-        &:focus {
-          outline: none;
-        }
-      }
-
-      .commentReply {
-        position: absolute;
-        top: 0;
-        right: 6px;
-        bottom: 0;
-        width: 52px;
-        border-style: none;
-        background-color: white;
-        color: #0099ff;
-
-        &:disabled {
-          color: rgb(173, 214, 230);
-        }
-      }
-    }
-  }
+  float: left;
 }
