+++ src/pages/hyominShin/Main/MainRight/MainRight.scss
@@ -2,7 +2,7 @@
   flex-direction: column;
   width: 30%;
   height: 700px;
-  margin-top: 100px;
+  margin-top: 60px;
   margin-left: 20px;
   float: left;
