+++ src/pages/hyominShin/Main/MainShin.js
@@ -1,6 +1,6 @@
 /* eslint-disabled  */
 
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import Nav from './nav';
 import './MainShin.scss';
 import '../../../styles/common.scss';
@@ -8,33 +8,29 @@ import ArticleBody from './Article/ArticleBody';
 import MainRight from './MainRight/MainRight';
 
 const Main = () => {
-  // const [commentObject, setcommentObject] = useState({
-  //   comment: commentArray,
-  //   heart: like,
-  // });
+  const [userList, setUserList] = useState([]);
+  // const {id, nickName, subText, profileImg, feedImg, comments, likes} = userList;
+  // 그 뭐지 객체구조 분해 할당 한거임
+
+  // MockData 불러온거
+  useEffect(() => {
+    fetch('/data/userData.json')
+      .then(response => response.json())
+      .then(userData => setUserList(userData));
+  }, []);
 
   return (
     <>
       <Nav />
       <div className="mainBody">
         {/* <!-- 메인 피드 --> */}
-        <ArticleBody />
+        <ArticleBody user={userList} />
+
         {/* <!-- 메인 오른쪽, 스토리와 연관사람 추천--> */}
         <MainRight />
       </div>
     </>
   );
 };
-
-// function Comment(props) {
-//   return(
-//     <li className="comment">
-//       <span className="textBold">someone</span>
-//       <span> { props.comment } </span>
-//       <img className="smallHeart" src="images/heart.png" />
-//       <span className="miniSubGray">댓글삭제</span>
-//     </li>
-//   )
-// }
 
 export default Main;
