+++ src/pages/hyominShin/Main/NavProfile/NavProfile.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import './NavProfile.scss';
+import '../../../../styles/common.scss';
+
+const NavProfile = () => {
+  const [isClicked, satIsClicked] = useState(false);
+  return <div className="" />;
+};
+
+export default NavProfile;
